+++ server/src/main/java/com/springboot/server/entities/Print.java
@@ -15,15 +15,15 @@ import javax.persistence.*;
 @Table(name = "print")
 public class Print {
     @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-    private Integer id;
-    @Column(name ="owner",length = 100)
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+    @Column(name = "owner", length = 100)
     private String owner;
-    @Column(name ="namedocument",length = 100)
+    @Column(name = "namedocument", length = 100)
     private String nameDocument;
-    @Column(name ="numberpages")
+    @Column(name = "numberpages")
     private Integer numberPages;
-    @Column(name ="dateprint", length = 100)
+    @Column(name = "dateprint", length = 100)
     private String datePrint;
 
 }
